Rename CoinAudits module package
