+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/coin/audits/CoinAuditsModule.kt
@@ -1,4 +1,4 @@
-package io.horizontalsystems.bankwallet.modules.coin.audits_new
+package io.horizontalsystems.bankwallet.modules.coin.audits
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
